sore/sore/asdfghjkl;ASDFGHJKL/telemetry_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0,<3.0
aiokafka>=0.8.0
pyJWT>=2.8.0
python-dotenv>=1.0.0
structlog>=24.0.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=6.1.0
pytest>=8.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
